Add a Unit Frame anchor
Imagine a thin box in the center of the screen, which the frame are
attached to. Changing the width of the box to change the gap between
player and target. Castbars are attached to the top and bottom of the
frame. Party and Raid frames are also attached to this frame.