Switch to Ace3 as a base
This move a lot of code out of the core that we no longer need to
maintain because its the Framework that's doing it now.

We also need to temporarly disable the options until we rewrite them in
the Ace3 Framework

Removed test command as well as it is so buggy. But the code can stay.