Combine Update and Configure methods
There is no reason to have them seperate. When we update them, we want to update all the data, from the header attributes, to the style. (well, we could keep em seperate, but the 'Update' method should do everything)

In commit 90eab4e i was having a issue of the update method being called multiple times (i guess. it was 2 years ago). I think was really due to the oUF init callback. So now we can just call the update method after creation. It will trickel down to all the oUF objects.